Refactoring to use strategy pattern with game dynamics in a module, included in modules of the game versions and these included when Game class is instanciated
