+++ views/index.erb
@@ -1,12 +1,5 @@
 <h1>Rock Paper Scissors</h1>
 
-<h2>Rules</h2>
-<ul>
-  <% (Game::RulesOptions::rules).each do |rule| %>
-    <li><%= rule %></li>
-  <% end %>
-</ul>
-
 <form action='/start'>
 
   <p>Select the level of game you want to play</p>
